areaselector has been moved out from toolbar
