+++ src/client/components/PageLayout/Toolbar/Toolbar.scss
@@ -32,76 +32,6 @@
   font-weight: 400;
 }
 
-.country-select__select-label {
-  display: none;
-  font-size: $font-l;
-  margin-right: 32px;
-
-  @include min-width($laptop) {
-    display: initial;
-  }
-}
-
-.btn-country-select {
-  position: relative;
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  min-width: 250px;
-  padding: 4px 12px;
-  background-color: $ui-bg;
-  border: none;
-  text-align: left;
-  font-size: $font-m;
-  font-weight: normal;
-  font-family: inherit;
-  color: $text-body;
-  transition: $ease-out;
-  outline: unset;
-
-  &:hover {
-    background-color: darken($ui-bg, 4%);
-  }
-
-  .icon {
-    margin-right: unset;
-  }
-}
-
-.toolbar__country {
-  display: grid;
-  grid-template-columns: 18px 1fr;
-  align-items: center;
-  color: $text-body;
-  width: 100%;
-
-  .flag {
-    background-repeat: no-repeat;
-    background-size: cover;
-    width: 18px;
-    height: 18px;
-    box-shadow: inset 0 0 0 1px rgba(0, 0, 0, 0.15);
-    border-radius: 2px;
-  }
-
-  .name-container {
-    padding-left: $spacing-xxs;
-  }
-
-  .name {
-    font-weight: 600;
-    @include line-clamp(1);
-  }
-
-  .user-role {
-    font-size: $font-xxs;
-    font-weight: 600;
-    color: $text-disabled;
-    margin-top: -3px;
-    opacity: 0.8;
-  }
-}
-
 .toolbar__select-laptop {
   display: none;
