+++ src/client/utils/index.ts
@@ -1,4 +1,5 @@
 export { Breakpoints } from './breakpoints'
+export { checkMatch } from './checkMatch'
 export { Dates } from './dates'
 export { MapController } from './mapController'
 export { Urls } from './urls'
